refactor: rust wasm lib for KCL
